refactor(common settings): extract parameter name control to separate component
