+++ src/ui/units/dash/containers/Dialogs/Control2/Sections/CommonSettingsSection/InputSettings/InputSettings.tsx
@@ -1,19 +1,13 @@
 import React from 'react';
 
-import {FormRow} from '@gravity-ui/components';
-import {TextInput} from '@gravity-ui/uikit';
-import {FieldWrapper} from 'components/FieldWrapper/FieldWrapper';
 import {I18n} from 'i18n';
-import {useDispatch, useSelector} from 'react-redux';
-import {DialogControlQa} from 'shared/constants/qa/control';
-import {setSelectorDialogItem} from 'units/dash/store/actions/dashTyped';
-import {selectSelectorDialog} from 'units/dash/store/selectors/dashTypedSelectors';
 
 import {SectionWrapper} from '../../../../../../../../components/SectionWrapper/SectionWrapper';
 import {OperationSelector} from '../../OperationSelector/OperationSelector';
 import {ValueSelector} from '../../ValueSelector/ValueSelector';
 import type {ValueSelectorControlProps} from '../../ValueSelector/types';
 import {InputTypeSelector} from '../InputTypeSelector/InputTypeSelector';
+import {ParameterNameInput} from '../ParameterNameInput/ParameterNameInput';
 import {getElementOptions} from '../helpers/input-type-select';
 
 const i18n = I18n.keyset('dash.control-dialog.edit');
@@ -23,19 +17,6 @@ type InputSettingsProps = {
 };
 
 const InputSettings = ({isSectionHidden}: InputSettingsProps) => {
-    const dispatch = useDispatch();
-    const {fieldName, validation} = useSelector(selectSelectorDialog);
-
-    const handleFieldNameUpdate = React.useCallback((value: string) => {
-        dispatch(
-            setSelectorDialogItem({
-                fieldName: value,
-            }),
-        );
-    }, []);
-
-    const value = fieldName ?? '';
-
     const options = React.useMemo(() => {
         return getElementOptions();
     }, []);
@@ -51,15 +32,7 @@ const InputSettings = ({isSectionHidden}: InputSettingsProps) => {
                 isStylesHidden={isSectionHidden}
                 title={isSectionHidden ? '' : i18n('label_common-settings')}
             >
-                <FormRow label={i18n('field_field-name')}>
-                    <FieldWrapper error={validation.fieldName}>
-                        <TextInput
-                            qa={DialogControlQa.fieldNameInput}
-                            value={value}
-                            onUpdate={handleFieldNameUpdate}
-                        />
-                    </FieldWrapper>
-                </FormRow>
+                <ParameterNameInput label={i18n('field_field-name')} />
                 <InputTypeSelector options={options} />
                 <OperationSelector />
                 <ValueSelector controlProps={controlProps} />
